% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Desc.ts.R
\name{plot.Desc.ts}
\alias{plot.Desc.ts}
\title{Plot frequency tables with bar or dot representation}
\usage{
\method{plot}{Desc.ts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Desc.table}.}

\item{...}{Graphical parameters passed to \code{par()}, such as
\code{cex}, \code{las}, \code{mar}, or \code{oma}.}
}
\description{
Creates a ACF plot for objects produced by \code{Desc.ts()}.
Two mosaicplots are placed side by side
}
\details{
Graphical parameters supplied via \code{...} are applied globally using
\code{par()} and affect all subplots consistently.
}
