% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorr.R
\name{plotCorr}
\alias{plotCorr}
\title{Plot a Correlation Matrix}
\usage{
plotCorr(
  x,
  cols = colorRampPalette(c(Pal()[2], "white", Pal()[1]), space = "rgb")(20),
  breaks = seq(-1, 1, length = length(cols) + 1),
  border = "grey",
  lwd = 1,
  args.colorLegend = NULL,
  xaxt = par("xaxt"),
  yaxt = par("yaxt"),
  cex.axis = 0.8,
  las = 2,
  mar = c(3, 8, 8, 8),
  mincor = 0,
  main = "",
  clust = FALSE,
  ...
)
}
\arguments{
\item{x}{x is a correlation matrix to be visualized.}

\item{cols}{the colors for shading the matrix. Uses the package's option
\code{"col1"} and \code{"col2"} as default.}

\item{breaks}{a set of breakpoints for the colours: must give one more
breakpoint than colour. These are passed to \code{image()} function.  If
breaks is specified then the algorithm used follows \code{\link{cut}}, so
intervals are closed on the right and open on the left except for the lowest
interval.}

\item{border}{color for borders. The default is \code{grey}. Set this
argument to \code{NA} if borders should be omitted.}

\item{lwd}{line width for borders. Default is 1.}

\item{args.colorLegend}{list of arguments for the \code{\link{colorLegend}}.
Use \code{NA} if no color legend should be painted.}

\item{xaxt}{parameter to define, whether to draw an x-axis, defaults to
\code{"n"}.}

\item{yaxt}{parameter to define, whether to draw an y-axis, defaults to
\code{"n"}.}

\item{cex.axis}{character extension for the axis labels.}

\item{las}{the style of axis labels.}

\item{mar}{sets the margins, defaults to mar = c(3, 8, 8, 8) as we need a
bit more room on the right.}

\item{mincor}{numeric value between 0 and 1, defining the smallest
correlation that is to be displayed. If this is >0 then all correlations
with a lower value are suppressed.}

\item{main}{character, the main title.}

\item{clust}{logical. If set to \code{TRUE}, the correlations will be
clustered in order to aggregate similar values.}

\item{\dots}{the dots are passed to the function \code{\link{image}}, which
produces the plot.}
}
\value{
no values returned.
}
\description{
This function produces a graphical display of a correlation matrix. The
cells of the matrix can be shaded or colored to show the correlation value.
}
\examples{

m <- cor(swiss)

plotCorr(m, cols=colorRampPalette(c("red", "black", "green"), space = "rgb")(20))
plotCorr(m, cols=colorRampPalette(c("red", "black", "green"), space = "rgb")(20),
         args.colorLegend=NA)

m <- cor(mtcars)
plotCorr(m, col=Pal("RedWhiteBlue1", 100), border="grey",
         args.colorLegend=list(labels=format(seq(-1,1,.25), digits=2), frame="grey"))

# display only correlation with a value > 0.7
plotCorr(m, mincor = 0.7)
x <- matrix(rep(1:ncol(m),each=ncol(m)), ncol=ncol(m))
y <- matrix(rep(ncol(m):1,ncol(m)), ncol=ncol(m))
txt <- format(m, digits=3)
idx <- upper.tri(matrix(x, ncol=ncol(m)), diag=FALSE)

# place the text on the upper triagonal matrix
text(x=x[idx], y=y[idx], label=txt[idx], cex=0.8, xpd=TRUE)

# put similiar correlations together
plotCorr(m, clust=TRUE)

# same as
idx <- order.dendrogram(as.dendrogram(
          hclust(dist(m), method = "mcquitty")
       ))
plotCorr(m[idx, idx])

# plot only upper triangular matrix and move legend to bottom
m <- cor(mtcars)
m[lower.tri(m, diag=TRUE)] <- NA

# get the p-values
p <- outer(
  (vars <- colnames(mtcars)), vars,
  Vectorize(function(v1, v2)
    cor.test(mtcars[[v1]], mtcars[[v2]], method = "pearson")$p.value
  )
)
dimnames(p) <- list(vars, vars)
m[p > 0.05] <- NA

plotCorr(m, mar=c(8,8,8,8), yaxt="n",
         args.colorLegend = list(x="bottom", inset=-.15, horiz=TRUE, 
                                 height=abs(lineToUser(line = 2.5, side = 1)), 
                                 width=ncol(m)))
mtext(text = rev(rownames(m)), side = 4, at=1:ncol(m), las=1, line = -5, cex=0.8)

}
\seealso{
\code{\link{image}}, \code{\link{colorLegend}},
\code{\link[corrgram]{corrgram}()}, \code{\link{plotWeb}()}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{hplot}
\keyword{multivariate}
